+++ scripts/prepare-release.sh
@@ -44,8 +44,9 @@ npm prune --production
 rm -rf .release
 mkdir -p .release
 
-cp action.yml package.json package-lock.json .release/
-rsync -R dist/src/*.js .release/
+cp action.yml action-types.yml package.json package-lock.json .release/
+rsync -R -v dist/src/*.js .release/
+rsync -R -v dist/src/**/*.js .release/
 cp -R node_modules .release/node_modules
 
 git checkout "$version"
@@ -57,11 +58,12 @@ rm -rf dist
 mkdir -p dist/src
 
 mv .release/action.yml .
+mv .release/action-types.yml .
 mv .release/dist/src/ ./dist/
 mv .release/*.json .
 mv .release/node_modules .
 
-git add action.yml ./dist/src/*.js package.json package-lock.json node_modules
+git add action.yml action-types.yml ./dist/src/*.js package.json package-lock.json node_modules
 set +x
 
 echo "Done. Please check 'git diff --cached' to verify changes. If ok, add version tag and push it to remote"