+++ test/fakedOctokit.ts
@@ -1,0 +1,34 @@
+type OctokitOpts = { owner: string; repo: string; commit_sha: string; body: string };
+class FakedOctokitRepos {
+    spyOpts: OctokitOpts[];
+    constructor() {
+        this.spyOpts = [];
+    }
+    createCommitComment(opt: OctokitOpts) {
+        this.spyOpts.push(opt);
+        return Promise.resolve({
+            status: 201,
+            data: {
+                html_url: 'https://dummy-comment-url',
+            },
+        });
+    }
+    lastCall(): OctokitOpts {
+        return this.spyOpts[this.spyOpts.length - 1];
+    }
+    clear() {
+        this.spyOpts = [];
+    }
+}
+
+export const fakedRepos = new FakedOctokitRepos();
+
+export class FakedOctokit {
+    rest = {
+        repos: fakedRepos,
+    };
+    opt: { token: string };
+    constructor(token: string) {
+        this.opt = { token };
+    }
+}